+++ .rokct/cache/delivery/templates/pages/driver/order_history/order_history.dart
@@ -124,7 +124,7 @@ class _OrderHistoryPageState extends ConsumerState<OrderHistoryPage> {
                     isDarkMode: true);
               },
               child: Container(
-                decoration: const BoxDecoration(
+                decoration: BoxDecoration(
                   shape: BoxShape.circle,
                   color: AppStyle.primary,
                 ),