+++ .rokct/cache/delivery/templates/pages/driver/parcel_history/parcel_history.dart
@@ -129,7 +129,7 @@ class _ParcelHistoryPageState extends ConsumerState<ParcelHistoryPage> {
                     isDarkMode: true);
               },
               child: Container(
-                decoration: const BoxDecoration(
+                decoration: BoxDecoration(
                   shape: BoxShape.circle,
                   color: AppStyle.primary,
                 ),