+++ .rokct/cache/delivery/templates/pages/driver/profile/widgets/edit_profile_modal.dart
@@ -309,7 +309,6 @@ class _EditProfileModalState extends ConsumerState<EditProfileModal> {
                           context: context,
                           modal: const EditCar(),
                           isDarkMode: false,
-                          isExpanded: true,
                         );
                       },
                       child: Container(